Fix clippy warnings and formatting issues
- Fix std::io::Error::new to use std::io::Error::other
- Fix large enum variant by boxing TLS stream
- Add type aliases to reduce type complexity
- Fix redundant closures and map_or patterns
- Fix needless borrows in slice comparisons
- Fix format! in format args patterns
- Add allow attributes for test-specific warnings
- Fix collapsible if statements
- Run cargo fmt for consistent formatting

+++ crates/kafka-backup-core/tests/unit_suite/helpers.rs
@@ -93,7 +93,7 @@ pub fn create_segment_metadata(
         end_offset,
         start_timestamp,
         end_timestamp,
-        record_count: (end_offset - start_offset + 1) as i64,
+        record_count: end_offset - start_offset + 1,
         compressed_size: 1024,
         uncompressed_size: 4096,
     }
